fix: View -> input & output 분리
